+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Updates
 
+- Updates the `FilterBarInline` component to apply `closeOnBlur` to `MultiSelect` components when `layout="row"`.
 - Updates `Breadcrumbs` component props to include `customLinkComponent` and `linkProps`.
 - Updates Banner component to allow for HTML content in the `content` prop when passed as a string.
 - Updates `Searchbar` component styles on mobile breakpoints
@@ -19,6 +20,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Adds
 
 - Adds the `FilterBarInline` component.
+- Adds the `FilterBarPopup` compoonent and `useFilterBarPopup` hook.
 - Adds refined `NewsletterSignup` component with updated props.
 - Adds refined `SocialMediaLinks` component with updated props